requirements.txt
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
pydantic==2.10.4
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
werkzeug==3.0.6
zipp==3.20.2
